requirements.txt
# ── Core API ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0          # async SQLite driver

# ── Task Queue ────────────────────────────────────────────
celery[redis]==5.3.6
redis==5.0.4

# ── Scraping ──────────────────────────────────────────────
playwright==1.44.0
httpx==0.27.0
dnspython==2.6.1

# ── AI ────────────────────────────────────────────────────
google-generativeai==0.7.2

# ── Report Generation ─────────────────────────────────────
jinja2==3.1.4
weasyprint==62.3          # PDF export (Linux/macOS)
# Alternative for Windows — uncomment if WeasyPrint fails:
# pyppeteer==2.0.0        # Chromium-based PDF (fallback)

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
Pillow==10.3.0
tldextract==5.1.2
python-whois==0.9.4
cryptography==42.0.7      # SSL cert analysis
pydantic-settings==2.2.1
aiofiles==23.2.1
